backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and storage
sqlalchemy==2.0.23
alembic==1.12.1

# Vector database
chromadb==0.4.18

# RAG and AI processing
raganything>=0.1.0
lightrag-hku>=0.1.0
mineru>=0.2.0

# OpenAI API (user-configurable)
openai>=1.0.0

# Image processing (for MinerU and multimodal)
Pillow>=10.0.0

# PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.9.0

# Audio processing
mutagen>=1.47.0

# Video processing (optional - requires system ffmpeg)
# ffmpeg-python>=0.2.0

# Async task processing
celery==5.3.4
redis==5.0.1

# Data validation and serialization
pydantic>=2.7.1,<3.0.0
pydantic-settings>=2.2.1,<3.0.0

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Graph processing
networkx==3.2.1

# PyTorch with CUDA support for MinerU acceleration
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1

# Numpy compatibility fix for ChromaDB
numpy>=1.22.5,<2.0.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0